refactor : 로또 구입 금맥 및 구매 가능 횟수 타입 변경

+++ src/main/java/lotto/domain/Lottos.java
@@ -12,13 +12,13 @@ public class Lottos {
     private static final int LOTTO_SIZE=6;
 
 
-    public Lottos(int count){
+    public Lottos(long count){
     lottos = new ArrayList<>();
     generateRandomNumber(count);
 
 }
 
-    private void generateRandomNumber(int count) {
+    private void generateRandomNumber(long count) {
         for (int i = 0; i < count; i++) {
             List<Integer> numbers = Randoms.pickUniqueNumbersInRange(
                     MIN_LOTTO_NUMBER,
